fixed 分享领券、平台方权限控制

+++ backend/fuint-application/src/main/java/com/fuint/module/backendApi/controller/BackendUserGradeController.java
@@ -163,6 +163,11 @@ public class BackendUserGradeController extends BaseController {
             return getFailureResult(1001, "请先登录");
         }
 
+        MtUserGrade mtUserGrade = userGradeService.queryUserGradeById(0, id, 0);
+        if (mtUserGrade == null || !mtUserGrade.getMerchantId().equals(accountInfo.getMerchantId())) {
+            return getFailureResult(201, "您没有删除权限");
+        }
+
         String operator = accountInfo.getAccountName();
         userGradeService.deleteUserGrade(id, operator);
 
@@ -184,6 +189,10 @@ public class BackendUserGradeController extends BaseController {
         AccountInfo accountInfo = TokenUtil.getAccountInfoByToken(token);
         if (accountInfo == null) {
             return getFailureResult(1001, "请先登录");
+        }
+
+        if (accountInfo.getMerchantId() == null || accountInfo.getMerchantId() <= 0) {
+            return getFailureResult(201, "平台方帐号无法执行该操作，请使用商户帐号操作");
         }
 
         String grade = param.get("grade") == null ? "0" : param.get("grade").toString();
